refactor(*): GetPurchaseResponse 위치 및 네이밍 변경
