+++ shared/index.ts
@@ -1,5 +1,4 @@
 export * from './constants/category';
 export * from './schemas/purchase.schema';
 export * from './types/draft';
-export * from './types/purchase.type';
 export * from './types/response.type';
