+++ shared/types/response.type.ts
@@ -1,5 +1,42 @@
 import { Category } from '../constants/category';
 
+// TODO: 사입내역 추가 API 응답 추가
+
+export interface PurchaseRecord {
+  purchaseId: string;
+  productId: string;
+  purchaseNo: string;
+  productNo: string;
+  purchasedAt: string;
+  vendor: string;
+  product: string;
+  category: Category;
+  color: string | null;
+  size: string | null;
+  option: string | null;
+  price: number;
+  quantity: number;
+  totalPrice: number;
+  backorderQuantity: number;
+  receipt: string | null;
+}
+
+export interface PaginationMeta {
+  total: number;
+  page: number;
+  limit: number;
+  totalPages: number;
+}
+
+// 사입내역 조회 API 응답
+export interface GetPurchaseResponse {
+  success: boolean;
+  status: number;
+  message: string;
+  records: PurchaseRecord[];
+  pagination: PaginationMeta;
+}
+
 export interface PurchaseDetailProduct {
   productId: string;
   productNo: string;
